# app/apirouter.py
from contextlib import asynccontextmanager
import logging
from fastapi import FastAPI, Query, HTTPException, Header, Depends, Request
from fastapi.responses import RedirectResponse, FileResponse
from typing import Literal
from dotenv import load_dotenv
import os

from pybeansack.models import *
from app.shared import AppContext
from app.shared.consts import *

load_dotenv()

### App Info ###
NAME = "Beans API & MCP"
DESCRIPTION = "Beans is an intelligent news & blogs aggregation and search service that curates fresh content from RSS feeds using AI-powered natural language queries and filters. Access trending articles, publishers, categories, and entities with fast JSON responses, perfect for developers building news apps or integrating smart content discovery."
VERSION = "0.1"
FAVICON = "app/assets/images/beans.png"

### DB QUERY CONSTANTS ###
DEFAULT_ACCURACY = 0.75
DEFAULT_LIMIT = 16

PROCESSED_ITEMS = ["gist IS NOT NULL", "embedding IS NOT NULL"]
UNRESTRICTED_CONTENT = ["restricted_content IS NULL", "content IS NOT NULL"]
CORE_BEAN_FIELDS = [
    K_URL, K_KIND,
    K_TITLE, K_SUMMARY,
    K_AUTHOR, K_SOURCE, K_IMAGEURL, K_CREATED, 
    K_CATEGORIES, K_SENTIMENTS, K_REGIONS, K_ENTITIES
]
EXTENDED_BEAN_FIELDS = CORE_BEAN_FIELDS + [K_CONTENT]
CORE_PUBLISHER_FIELDS = [K_SOURCE, K_BASE_URL, K_SITE_NAME, K_DESCRIPTION, K_FAVICON]

### QUERY PARAMETER DEFINITIONS ###
Q = Query(
    default=None, 
    min_length=3, 
    max_length=512,
    description="The search query used for vector search. Minimum length is 3 characters."
)
SEARCH_TYPE = Query(
    default="vector", 
    description="Indicate if the search should be a semantic (vector) search or a text keyword (bm25) search."
)
ACCURACY = Query(
    default=DEFAULT_ACCURACY, 
    ge=0, le=1,
    description="""Cosine similarity score for vector search. Includes articles with a `cosine_similarity_score >= accuracy` in search result. The higher the accuracy the stricter the filter, meaning less items will be returned. `accuracy=1` means only exactly identical articles, `accuracy=0` means all articles regardless of similarity will be returned."""
)
URL = Query(
    ..., 
    min_length=10, 
    description="The URL of the article for which related articles are to be found. Minimum length is 10 characters."
)
KIND = Query(
    default=None, 
    description="Type of article to query: `news`, `blog`, or keep it unspecified for both."
)
CATEGORIES = Query(
    max_length=MAX_LIMIT, 
    default=None, 
    description="Filters articles by one or more categories (case sensitive). Used for faster keyword search."
)
ENTITIES = Query(
    max_length=MAX_LIMIT, 
    default=None, 
    description="Filters articles by one or more named entities (case sensitive). Used for faster keyword search."
)
REGIONS = Query(
    max_length=MAX_LIMIT, 
    default=None, 
    description="Filters articles by one or more regions (case sensitive). Used for faster keyword search."
)
TAGS = Query(
    max_length=MAX_LIMIT, 
    default=None, 
    description="Filters articles by one or more tags (combining categories, regions, and entities). Ignores case and non-alphanumeric characters, so `Data Protection`, `data PRotecTION`, `data-protection`,  and `#data-protection` are treated identically. All specified tags must match for an item to be included in results. Provides a fuzzier, more fault-tolerant alternative to case-sensitive category, region, and entity filters.",
    examples=["""["Artificial Intelligence", "UK", "saudi-arabia", "Elon-Musk", "#data-protection"]"""]
)
SOURCES = Query(
    max_length=MAX_LIMIT, 
    default=None, 
    description="Filters articles by one or more publisher IDs (case sensitive)."
)
AUTHORS = Query(
    max_length=MAX_LIMIT, 
    default=None, 
    description="Filters articles by one or more authors (case sensitive)."
)
PUBLISHED_SINCE = Query(
    default=None, 
    description="Filters articles that were published on or after this datetime."
)
TRENDING_SINCE = Query(
    default=None, 
    description="Filters articles that had social media engagement (comments, shares, likes) on or after this datetime."
)
WITH_CONTENT = Query(
    default=False, 
    description="""Includes full text content (plaintext/markdown) of the articles. This applies ONLY to articles where full content is available. By default (when `with_content=false`), ONLY `summary` is included instead of `content` for faster response."""
)
GROUP_BY = Query(
    default=None, 
    description="Return one item per `source`, `author` or `cluster` (news and blogs about the same content are grouped in one cluster)."
)
OFFSET = Query(
    ge=0, 
    default=0, 
    description="Number of items to skip (for pagination)."
)
LIMIT = Query(
    ge=MIN_LIMIT, 
    le=MAX_LIMIT, 
    default=DEFAULT_LIMIT, 
    description="Maximum number of items to return."
)

db_context: AppContext = None

### ROUTER AND ROUTE DEFINITIONS ###
@asynccontextmanager
async def lifespan(app: FastAPI):
    global db_context
    
    # Load API keys configuration
    api_keys = os.getenv("API_KEYS")
    if api_keys:
        api_keys = api_keys.split(";")
        api_keys = [kv.strip().split("=", maxsplit=1) for kv in api_keys if kv.strip()]
        api_keys = {kv[0].strip(): kv[1].strip() for kv in api_keys}
    
    db_context = AppContext(
        db_kwargs={
            "db_type": os.getenv("DB_TYPE"), 
            "pg_connection_string": os.getenv("PG_CONNECTION_STRING")
        },
        embedder_kwargs={
            "model_name": os.getenv('EMBEDDER_MODEL'), 
            "ctx_len": int(os.getenv('EMBEDDER_CTX', 512))
        },
        api_keys=api_keys
    )    
    yield    

    db_context.close()

def verify_api_key(request: Request):
    # Check each allowed header in the request
    api_keys = db_context.settings.get('api_keys')
    if not api_keys: return True # If no API keys are configured, allow all requests
    if any(request.headers.get(header) == value for header, value in api_keys.items()): return True
    raise HTTPException(status_code=401, detail="Invalid API Key")

api_key_dependency = Depends(verify_api_key)

app = FastAPI(title=NAME, version=VERSION, description=DESCRIPTION, lifespan=lifespan)

# @app.get("/")
# async def root():
#     return FileResponse("app/assets/index.html")

@app.get(
    "/health", 
    summary="Health check", 
    description="Performs a health check to see if the API service is running. Returns `{\"status\": \"alive\"}` if the service is operational."
)
async def health_check():
    return {"status": "alive"}

@app.get(
    "/favicon.ico", 
    summary="Get favicon", 
    description="Retrieves the favicon for the API service."
)
async def get_favicon():
    return FileResponse(FAVICON, media_type="image/png")

@app.get(
    "/tags/categories", 
    summary="List categories",
    dependencies=[api_key_dependency],
    description="Retrieves a list of unique values of articles categories/topics, such as Artificial Intelligence, Cybersecurity, Politics, Software Engineering etc."
)
async def get_categories(offset: int = OFFSET, limit: int = LIMIT) -> list[str]:
    return db_context.db.distinct_categories(limit=limit, offset=offset)

@app.get(
    "/tags/entities", 
    summary="List entities",
    dependencies=[api_key_dependency],
    description="Retrieves a list of unique values of named entities (people, organizations, products) mentioned in the articles."
)
async def get_entities(offset: int = OFFSET, limit: int = LIMIT) -> list[str]:
    return db_context.db.distinct_entities(limit=limit, offset=offset)

@app.get(
    "/tags/regions", 
    summary="List regions",
    dependencies=[api_key_dependency],
    description="Retrieves a list of unique values of geographic regions mentioned in the articles such as UK, US, Europe etc."
)
async def get_regions(offset: int = OFFSET, limit: int = LIMIT) -> list[str]:
    return db_context.db.distinct_regions(limit=limit, offset=offset)

@app.get(
    "/articles/latest", 
    summary="Search latest articles",
    dependencies=[Depends(verify_api_key)],
    response_model_exclude_none=True,
    response_model_exclude_unset=True,
    response_model_exclude=[K_COLLECTED, K_RESTRICTED_CONTENT, "title_length", "summary_length", "content_length", K_EMBEDDING, K_GIST],
    description="""Searches for the latest articles/news/blogs. For vector search (when `q` is provided), the results are sorted by relevance. Otherwise, they are sorted by publication date in descending order (newest first)."""
)
async def get_latest_articles(
    q: str = Q,
    acc: float = ACCURACY,
    kind: Literal[NEWS, BLOG] = KIND,
    # categories: list[str] = CATEGORIES,
    # entities: list[str] = ENTITIES,
    # regions: list[str] = REGIONS,
    tags: list[str] = TAGS,
    sources: list[str] = SOURCES,
    published_since: datetime = PUBLISHED_SINCE,
    with_content: bool = WITH_CONTENT,
    limit: int = LIMIT,
    offset: int = OFFSET,
) -> Optional[list[Bean]]:    
    embedding = db_context.embed_query(q) if q else None
    distance = 1 - acc if q else 0
    return db_context.db.query_latest_beans(
        kind=kind,
        created=published_since,
        # categories=categories,
        # regions=regions,
        # entities=entities,
        tags=tags,
        sources=sources,
        embedding=embedding,
        distance=distance,
        conditions=UNRESTRICTED_CONTENT+PROCESSED_ITEMS if with_content else PROCESSED_ITEMS,
        limit=limit,
        offset=offset,        
        columns=EXTENDED_BEAN_FIELDS if with_content else CORE_BEAN_FIELDS
    )

@app.get(
    "/articles/trending", 
    summary="Search trending articles",
    dependencies=[Depends(verify_api_key)],
    response_model_exclude_none=True,
    response_model_exclude_unset=True,
    response_model_exclude=[K_COLLECTED, K_RESTRICTED_CONTENT, "title_length", "summary_length", "content_length", K_EMBEDDING, K_GIST],
    description="""Searches for the trending articles/news/blogs. For vector search (when `q` is provided), the results are sorted by relevance. Otherwise, they are sorted by internal trend score (calculated from social media engagement: comments, likes, shares, last engagement etc.)."""
)
async def get_trending_articles(
    q: str = Q,
    acc: float = ACCURACY,
    kind: Literal[NEWS, BLOG] = KIND,
    # categories: list[str] = CATEGORIES,
    # entities: list[str] = ENTITIES,
    # regions: list[str] = REGIONS,
    tags: list[str] = TAGS,
    sources: list[str] = SOURCES,
    trending_since: datetime = TRENDING_SINCE,
    with_content: bool = WITH_CONTENT,    
    limit: int = LIMIT,
    offset: int = OFFSET
) -> Optional[list[Bean]]:    
    embedding = db_context.embed_query(q) if q else None
    distance = 1 - acc if q else 0
    return db_context.db.query_trending_beans(
        kind=kind,
        updated=trending_since,
        # categories=categories,
        # regions=regions,
        # entities=entities,
        tags=tags,
        sources=sources,
        embedding=embedding,
        distance=distance,
        conditions=UNRESTRICTED_CONTENT+PROCESSED_ITEMS if with_content else PROCESSED_ITEMS,
        limit=limit,
        offset=offset,        
        columns=EXTENDED_BEAN_FIELDS if with_content else CORE_BEAN_FIELDS
    )

# TODO: add routes for /related

# @api.get(
#     "/related", 
#     response_model=list[Bean]|None,
#     response_model_by_alias=True,
#     response_model_exclude_none=True,
#     response_model_exclude_unset=True,
#     description="Retrieve beans related to the bean identified by the given URL, with optional filters. The result is sorted by newest first."
# )
# async def get_related(
#     url: str = URL,
#     kind: Literal["news", "blogs", None] = KIND,
#     categories: list[str] = CATEGORIES,
#     entities: list[str] = ENTITIES,
#     regions: list[str] = REGIONS,
#     sources: list[str] = SOURCES,
#     authors: list[str] = AUTHORS,
#     published_after: datetime = PUBLISHED_AFTER,
#     with_content: bool = WITH_CONTENT,
#     limit: int = LIMIT
# ) -> list[Bean]:
#     """
#     Retrieves the related beans to the given bean.
#     """  
#     filter = create_filter(kind, categories, entities, regions, sources, authors, published_after, with_content)
#     project = create_projection(with_content) 
#     return db.query_related_beans(url, filter=filter, sort_by=NEWEST, limit=limit, project = project)

@app.get(
    "/publishers", 
    summary="Get publishers' metadata",
    dependencies=[api_key_dependency],
    response_model_exclude_none=True,
    response_model_exclude_unset=True,
    response_model_exclude=[K_RSS_FEED, K_COLLECTED],
    description="Retrieves publisher metadata filtered by one or more publisher IDs."
)
async def get_publishers(
    sources: list[str] = Query(..., max_length=MAX_LIMIT, description="List of publisher IDs to search by (case sensitive)."),     
    limit: int = LIMIT,
    offset: int = OFFSET,
) -> Optional[list[Publisher]]:
    return db_context.db.query_publishers(sources=sources, limit=limit, offset=offset, columns=CORE_PUBLISHER_FIELDS)

@app.get(
    "/publishers/sources", 
    summary="List publisher IDs",
    dependencies=[api_key_dependency],
    description="Retrieves a list of unique values of publisher IDs from which the articles are sourced."
)
async def get_publishers(offset: int = OFFSET, limit: int = LIMIT) -> Optional[list[str]]:
    return db_context.db.distinct_publishers(limit=limit, offset=offset)