# app/assets/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Espresso News",
    "description": "Espresso News API is an intelligent news aggregation and search service that curates fresh content from RSS feeds using AI-powered natural language queries and filters. Access trending articles, publishers, categories, and entities with fast JSON responses, perfect for developers building news apps or integrating smart content discovery.",
    "version": "0.0.2"
  },
  "paths": {
...
    "/tags/categories": {
      "get": {
        "summary": "List categories",
        "description": "Retrieves a list of unique values of articles categories/topics, such as Artificial Intelligence, Cybersecurity, Politics, Software Engineering etc.",
        "operationId": "get_categories_categories_get",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip (for pagination).",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip (for pagination)."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum number of items to return.",
              "default": 16,
              "title": "Limit"
            },
            "description": "Maximum number of items to return."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Get Categories Categories Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tags/entities": {
      "get": {
        "summary": "List entities",
        "description": "Retrieves a list of unique values of named entities (people, organizations, products) mentioned in the articles.",
        "operationId": "get_entities_entities_get",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip (for pagination).",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip (for pagination)."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum number of items to return.",
              "default": 16,
              "title": "Limit"
            },
            "description": "Maximum number of items to return."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Get Entities Entities Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tags/regions": {
      "get": {
        "summary": "List regions",
        "description": "Retrieves a list of unique values of geographic regions mentioned in the articles such as UK, US, Europe etc.",
        "operationId": "get_regions_regions_get",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip (for pagination).",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip (for pagination)."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum number of items to return.",
              "default": 16,
              "title": "Limit"
            },
            "description": "Maximum number of items to return."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Get Regions Regions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/articles/latest": {
      "get": {
        "summary": "Search latest articles",
        "description": "Searches for the latest articles/news/blogs. For vector search (when `q` is provided), the results are sorted by relevance. Otherwise, they are sorted by publication date in descending order (newest first).",
        "operationId": "get_latest_articles_articles_latest_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 3,
              "maxLength": 512,
              "description": "The search query used for vector search. Minimum length is 3 characters.",
              "title": "Q"
            },
            "description": "The search query used for vector search. Minimum length is 3 characters."
          },
          {
            "name": "acc",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 1,
              "minimum": 0,
              "description": "Cosine similarity score for vector search. Includes articles with a cosine similarity score \u003E= accuracy in search result. The higher the accuracy the stricter the filter, meaning less items will be returned. `accuracy=1` means only exactly identical articles, `accuracy=0` means all articles regardless of similarity will be returned.",
              "default": 0.75,
              "title": "Acc"
            },
            "description": "Cosine similarity score for vector search. Includes articles with a cosine similarity score \u003E= accuracy in search result. The higher the accuracy the stricter the filter, meaning less items will be returned. `accuracy=1` means only exactly identical articles, `accuracy=0` means all articles regardless of similarity will be returned."
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "news",
                "blog"
              ],
              "type": "string",
              "description": "Type of article to query: 'news', 'blog', or keep it unspecified for both.",
              "title": "Kind"
            },
            "description": "Type of article to query: 'news', 'blog', or keep it unspecified for both."
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "maxItems": 100,
              "description": "Filters articles by one or more tags (combining categories, regions, and entities). Ignores case and non-alphanumeric characters, so 'Data Protection', 'data PRotecTION', 'data-protection',  and '#data-protection' are treated identically. All specified tags must match for an item to be included in results. Provides a fuzzier, more fault-tolerant alternative to case-sensitive category, region, and entity filters.",
              "examples": [
                "[\"Artificial Intelligence\", \"UK\", \"saudi-arabia\", \"Elon-Musk\", \"#data-protection\"]"
              ],
              "title": "Tags"
            },
            "description": "Filters articles by one or more tags (combining categories, regions, and entities). Ignores case and non-alphanumeric characters, so 'Data Protection', 'data PRotecTION', 'data-protection',  and '#data-protection' are treated identically. All specified tags must match for an item to be included in results. Provides a fuzzier, more fault-tolerant alternative to case-sensitive category, region, and entity filters."
          },
          {
            "name": "sources",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "maxItems": 100,
              "description": "Filters articles by one or more publisher IDs (case sensitive).",
              "title": "Sources"
            },
            "description": "Filters articles by one or more publisher IDs (case sensitive)."
          },
          {
            "name": "published_since",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time",
              "description": "Filters articles that were published on or after this datetime.",
              "title": "Published Since"
            },
            "description": "Filters articles that were published on or after this datetime."
          },
          {
            "name": "with_content",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Includes full text content (plaintext/markdown) of the articles. This applies ONLY to articles where full content is available. By default (when `with_content=False`), ONLY `summary` is included instead of `content` for faster response.",
              "default": false,
              "title": "With Content"
            },
            "description": "Includes full text content (plaintext/markdown) of the articles. This applies ONLY to articles where full content is available. By default (when `with_content=False`), ONLY `summary` is included instead of `content` for faster response."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip (for pagination).",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip (for pagination)."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum number of items to return.",
              "default": 16,
              "title": "Limit"
            },
            "description": "Maximum number of items to return."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Bean"
                      }
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Latest Articles Articles Latest Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/publishers": {
      "get": {
        "summary": "Get publishers' metadata",
        "description": "Retrieves publisher metadata filtered by one or more publisher IDs.",
        "operationId": "get_publishers_publishers_get",
        "parameters": [
          {
            "name": "sources",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "maxItems": 100,
              "description": "List of publisher IDs to search by (case sensitive).",
              "title": "Sources"
            },
            "description": "List of publisher IDs to search by (case sensitive)."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip (for pagination).",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip (for pagination)."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum number of items to return.",
              "default": 16,
              "title": "Limit"
            },
            "description": "Maximum number of items to return."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Publisher"
                      }
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Publishers Publishers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/publishers/sources": {
      "get": {
        "summary": "List publisher IDs",
        "description": "Retrieves a list of unique values of publisher IDs from which the articles are sourced.",
        "operationId": "get_publishers_publishers_sources_get",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip (for pagination).",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip (for pagination)."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum number of items to return.",
              "default": 16,
              "title": "Limit"
            },
            "description": "Maximum number of items to return."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Publishers Publishers Sources Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Publisher": {
        "properties": {
          "source": {
            "type": "string",
            "minLength": 1,
            "title": "Source",
            "description": "The publisher ID/domain name of the publisher. This matches the source field in Bean."
          },
          "base_url": {
            "type": "string",
            "minLength": 1,
            "title": "Base Url",
            "description": "The base URL of the publisher."
          },
          "site_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Name",
            "description": "The name of the site."
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "A description/details of the publisher."
          },
          "favicon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Favicon",
            "description": "The URL of the publisher's favicon."
          }
        },
        "type": "object",
        "required": [
          "source",
          "base_url"
        ],
        "title": "Publisher",
        "description": "Metadata of the website, publication or social medium from which an article or chatter is sourced."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}